factor CborReaderState to a separate file, add xml docs
