+++ src/libraries/System.Security.Cryptography.Encoding/tests/Cbor/CborReader.SkipValue.cs
@@ -124,7 +124,7 @@ namespace System.Formats.Cbor
 
                 case CborReaderState.Null:
                 case CborReaderState.Boolean:
-                case CborReaderState.SpecialValue:
+                case CborReaderState.SimpleValue:
                     ReadSimpleValue();
                     break;
